requirements.txt

python-dotenv==1.0.1
web3==6.15.1
eth-account==0.11.0
nextcord==2.6.0
python-telegram-bot==20.8
firebase-admin==6.4.0
plotly==5.19.0
kaleido==0.2.1  #image export
qrcode==7.4.2 # Generate address QR code
